@@ -3,24 +3,24 @@
   "version": "0.0.0",
   "private": true,
   "dependencies": {
-    "ace-builds": "^1.4.12",
-    "ansi-to-html": "^0.6.14",
-    "dompurify": "^2.2.2",
-    "lodash": "^4.17.20",
+    "ace-builds": "^1.4.13",
+    "ansi-to-html": "^0.7.2",
+    "dompurify": "^2.3.4",
+    "lodash": "^4.17.21",
     "prop-types": "^15.7.2",
-    "react": "^16.13.1",
-    "react-ace": "^9.1.4",
-    "react-dom": "^16.13.1",
-    "react-redux": "^7.2.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.3",
-    "react-virtualized": "^9.22.2",
-    "redux": "^4.0.5",
-    "xterm": "^4.9.0",
+    "react": "^17.0.2",
+    "react-ace": "^9.5.0",
+    "react-dom": "^17.0.2",
+    "react-redux": "^7.2.6",
+    "react-router-dom": "^6.1.1",
+    "react-scripts": "4.0.3",
+    "react-virtualized": "^9.22.3",
+    "redux": "^4.1.2",
+    "xterm": "^4.15.0",
     "xterm-addon-attach": "^0.6.0",
-    "xterm-addon-fit": "^0.4.0",
+    "xterm-addon-fit": "^0.5.0",
     "xterm-addon-web-links": "^0.4.0",
-    "yaml": "^1.10.0"
+    "yaml": "^1.10.2"
   },
   "devDependencies": {
     "@fortawesome/fontawesome-free": "^5.14.0",
