@@ -16,7 +16,7 @@
  */
 import React, {useEffect} from 'react';
 import {connect} from 'react-redux';
-import {BrowserRouter as Router, Route, Switch} from 'react-router-dom';
+import {BrowserRouter as Router, Route, Routes, useMatch} from 'react-router-dom';
 import apis from './apis';
 import crb from './clusterrolebindings';
 import cRoles from './clusterroles';
@@ -94,78 +94,78 @@ const App = ({dispatch}) => {
   }, []); // eslint-disable-line react-hooks/exhaustive-deps
   return (
       <Router>
-        <Switch>
-          <Route exact path='/' component={Home} />
-          <Route exact path='/clusterrolebindings' component={crb.ClusterRoleBindingsPage} />
-          <Route exact path='/clusterrolebindings/:uidOrName' component={crb.ClusterRoleBindingsDetailPage} />
-          <Route exact path='/clusterrolebindings/:uid/edit' component={crb.ClusterRoleBindingsEditPage} />
-          <Route exact path='/clusterroles' component={cRoles.ClusterRolesPage} />
-          <Route exact path='/clusterroles/:uidOrName' component={cRoles.ClusterRolesDetailPage} />
-          <Route exact path='/clusterroles/:uid/edit' component={cRoles.ClusterRolesEditPage} />
-          <Route exact path='/configmaps' component={configMaps.ConfigMapsPage} />
-          <Route exact path='/configmaps/:uid' component={configMaps.ConfigMapsDetailPage} />
-          <Route exact path='/configmaps/:uid/edit' component={configMaps.ConfigMapsEditPage} />
-          <Route exact path='/cronjobs' component={cj.CronJobsPage} />
-          <Route exact path='/cronjobs/:uid' component={cj.CronJobsDetailPage} />
-          <Route exact path='/cronjobs/:uid/edit' component={cj.CronJobsEditPage} />
-          <Route exact path='/customresourcedefinitions' component={crd.CustomResourceDefinitionsPage} />
-          <Route exact path='/customresourcedefinitions/:uid' component={crd.CustomResourceDefinitionsDetailPage} />
-          <Route exact path='/customresourcedefinitions/:uid/edit' component={crd.CustomResourceDefinitionsEditPage} />
-          <Route exact path='/daemonsets' component={ds.DaemonSetsPage} />
-          <Route exact path='/daemonsets/:uid' component={ds.DaemonSetsDetailPage} />
-          <Route exact path='/daemonsets/:uid/edit' component={ds.DaemonSetsEditPage} />
-          <Route exact path='/deploymentconfigs' component={dc.DeploymentConfigsPage} />
-          <Route exact path='/deploymentconfigs/:uid' component={dc.DeploymentConfigsDetailPage} />
-          <Route exact path='/deploymentconfigs/:uid/edit' component={dc.DeploymentConfigsEditPage} />
-          <Route exact path='/deployments' component={deployments.DeploymentsPage} />
-          <Route exact path='/deployments/:uid' component={deployments.DeploymentsDetailPage} />
-          <Route exact path='/deployments/:uid/edit' component={deployments.DeploymentsEditPage} />
-          <Route exact path='/horizontalpodautoscalers' component={hpa.HorizontalPodAutoscalersPage} />
-          <Route exact path='/horizontalpodautoscalers/:uid' component={hpa.HorizontalPodAutoscalersDetailPage} />
-          <Route exact path='/horizontalpodautoscalers/:uid/edit' component={hpa.HorizontalPodAutoscalersEditPage} />
-          <Route exact path='/ingresses' component={ingresses.IngressesPage} />
-          <Route exact path='/ingresses/:uid' component={ingresses.IngressesDetailPage} />
-          <Route exact path='/ingresses/:uid/edit' component={ingresses.IngressEditPage} />
-          <Route exact path='/jobs' component={jobs.JobsPage} />
-          <Route exact path='/jobs/:uid' component={jobs.JobsDetailPage} />
-          <Route exact path='/jobs/:uid/edit' component={jobs.JobsEditPage} />
-          <Route exact path='/namespaces' component={ns.NamespacesPage} />
-          <Route exact path='/namespaces/:uidOrName' component={ns.NamespacesDetailPage} />
-          <Route exact path='/nodes' component={nodes.NodesPage} />
-          <Route exact path='/nodes/:name' component={nodes.NodesDetailPage} />
-          <Route exact path='/nodes/:uid/edit' component={nodes.NodesEditPage} />
-          <Route exact path='/persistentvolumeclaims' component={pvc.PersistentVolumeClaimsPage} />
-          <Route exact path='/persistentvolumeclaims/:uid' component={pvc.PersistentVolumeClaimsDetailPage} />
-          <Route exact path='/persistentvolumeclaims/:uid/edit' component={pvc.PersistentVolumeClaimsEditPage} />
-          <Route exact path='/persistentvolumes' component={pv.PersistentVolumesPage} />
-          <Route exact path='/persistentvolumes/:uid' component={pv.PersistentVolumesDetailPage} />
-          <Route exact path='/persistentvolumes/:uid/edit' component={pv.PersistentVolumesEditPage} />
-          <Route exact path='/pods' component={pods.PodsPage} />
-          <Route exact path='/pods/:uid' component={pods.PodsDetailPage} />
-          <Route exact path='/pods/:uid/edit' component={pods.PodsEditPage} />
-          <Route exact path='/pods/:uid/exec' component={pods.PodsExecPage} />
-          <Route exact path='/pods/:uid/logs' component={pods.PodsLogsPage} />
-          <Route exact path='/replicationcontrollers' component={rc.ReplicationControllersPage} />
-          <Route exact path='/replicationcontrollers/:uid' component={rc.ReplicationControllersDetailPage} />
-          <Route exact path='/replicationcontrollers/:uid/edit' component={rc.ReplicationControllersEditPage} />
-          <Route exact path='/roles' component={roles.RolesPage} />
-          <Route exact path='/roles/:uid' component={roles.RolesDetailPage} />
-          <Route exact path='/roles/:uid/edit' component={roles.RolesEditPage} />
-          <Route exact path='/routes' component={routes.RoutesPage} />
-          <Route exact path='/routes/:uid' component={routes.RoutesDetailPage} />
-          <Route exact path='/routes/:uid/edit' component={routes.RoutesEditPage} />
-          <Route exact path='/search' component={search.SearchPage} />
-          <Route exact path='/secrets' component={secrets.SecretsPage} />
-          <Route exact path='/secrets/:uid' component={secrets.SecretsDetailPage} />
-          <Route exact path='/secrets/:uid/edit' component={secrets.SecretsEditPage} />
-          <Route exact path='/services' component={services.ServicesPage} />
-          <Route exact path='/services/:uid' component={services.ServicesDetailPage} />
-          <Route exact path='/services/:uid/edit' component={services.ServicesEditPage} />
-          <Route exact path='/statefulsets' component={sts.StatefulSetsPage} />
-          <Route exact path='/statefulsets/:uid' component={sts.StatefulSetsDetailPage} />
-          <Route exact path='/statefulsets/:uid/edit' component={sts.StatefulSetsEditPage} />
-          {/*<Route component={Error404Page} />*/}
-        </Switch>
+        <Routes>
+          <Route exact path='/' element={<Home/>} />
+          <Route exact path='/clusterrolebindings' element={<crb.ClusterRoleBindingsPage/>} />
+          <Route exact path='/clusterrolebindings/:uidOrName' element={<crb.ClusterRoleBindingsDetailPage/>} />
+          <Route exact path='/clusterrolebindings/:uid/edit' element={<crb.ClusterRoleBindingsEditPage/>} />
+          <Route exact path='/clusterroles' element={<cRoles.ClusterRolesPage/>} />
+          <Route exact path='/clusterroles/:uidOrName' element={<cRoles.ClusterRolesDetailPage/>} />
+          <Route exact path='/clusterroles/:uid/edit' element={<cRoles.ClusterRolesEditPage/>} />
+          <Route exact path='/configmaps' element={<configMaps.ConfigMapsPage/>} />
+          <Route exact path='/configmaps/:uid' element={<configMaps.ConfigMapsDetailPage/>} />
+          <Route exact path='/configmaps/:uid/edit' element={<configMaps.ConfigMapsEditPage/>} />
+          <Route exact path='/cronjobs' element={<cj.CronJobsPage/>} />
+          <Route exact path='/cronjobs/:uid' element={<cj.CronJobsDetailPage/>} />
+          <Route exact path='/cronjobs/:uid/edit' element={<cj.CronJobsEditPage/>} />
+          <Route exact path='/customresourcedefinitions' element={<crd.CustomResourceDefinitionsPage/>} />
+          <Route exact path='/customresourcedefinitions/:uid' element={<crd.CustomResourceDefinitionsDetailPage/>} />
+          <Route exact path='/customresourcedefinitions/:uid/edit' element={<crd.CustomResourceDefinitionsEditPage/>} />
+          <Route exact path='/daemonsets' element={<ds.DaemonSetsPage/>} />
+          <Route exact path='/daemonsets/:uid' element={<ds.DaemonSetsDetailPage/>} />
+          <Route exact path='/daemonsets/:uid/edit' element={<ds.DaemonSetsEditPage/>} />
+          <Route exact path='/deploymentconfigs' element={<dc.DeploymentConfigsPage/>} />
+          <Route exact path='/deploymentconfigs/:uid' element={<dc.DeploymentConfigsDetailPage/>} />
+          <Route exact path='/deploymentconfigs/:uid/edit' element={<dc.DeploymentConfigsEditPage/>} />
+          <Route exact path='/deployments' element={<deployments.DeploymentsPage/>} />
+          <Route exact path='/deployments/:uid' element={<deployments.DeploymentsDetailPage/>} />
+          <Route exact path='/deployments/:uid/edit' element={<deployments.DeploymentsEditPage/>} />
+          <Route exact path='/horizontalpodautoscalers' element={<hpa.HorizontalPodAutoscalersPage/>} />
+          <Route exact path='/horizontalpodautoscalers/:uid' element={<hpa.HorizontalPodAutoscalersDetailPage/>} />
+          <Route exact path='/horizontalpodautoscalers/:uid/edit' element={<hpa.HorizontalPodAutoscalersEditPage/>} />
+          <Route exact path='/ingresses' element={<ingresses.IngressesPage/>} />
+          <Route exact path='/ingresses/:uid' element={<ingresses.IngressesDetailPage/>} />
+          <Route exact path='/ingresses/:uid/edit' element={<ingresses.IngressEditPage/>} />
+          <Route exact path='/jobs' element={<jobs.JobsPage/>} />
+          <Route exact path='/jobs/:uid' element={<jobs.JobsDetailPage/>} />
+          <Route exact path='/jobs/:uid/edit' element={<jobs.JobsEditPage/>} />
+          <Route exact path='/namespaces' element={<ns.NamespacesPage/>} />
+          <Route exact path='/namespaces/:uidOrName' element={<ns.NamespacesDetailPage/>} />
+          <Route exact path='/nodes' element={<nodes.NodesPage/>} />
+          <Route exact path='/nodes/:name' element={<nodes.NodesDetailPage/>} />
+          <Route exact path='/nodes/:uid/edit' element={<nodes.NodesEditPage/>} />
+          <Route exact path='/persistentvolumeclaims' element={<pvc.PersistentVolumeClaimsPage/>} />
+          <Route exact path='/persistentvolumeclaims/:uid' element={<pvc.PersistentVolumeClaimsDetailPage/>} />
+          <Route exact path='/persistentvolumeclaims/:uid/edit' element={<pvc.PersistentVolumeClaimsEditPage/>} />
+          <Route exact path='/persistentvolumes' element={<pv.PersistentVolumesPage/>} />
+          <Route exact path='/persistentvolumes/:uid' element={<pv.PersistentVolumesDetailPage/>} />
+          <Route exact path='/persistentvolumes/:uid/edit' element={<pv.PersistentVolumesEditPage/>} />
+          <Route exact path='/pods' element={<pods.PodsPage/>} />
+          <Route exact path='/pods/:uid' element={<pods.PodsDetailPage/>} />
+          <Route exact path='/pods/:uid/edit' element={<pods.PodsEditPage/>} />
+          <Route exact path='/pods/:uid/exec' element={<pods.PodsExecPage/>} />
+          <Route exact path='/pods/:uid/logs' element={<pods.PodsLogsPage/>} />
+          <Route exact path='/replicationcontrollers' element={<rc.ReplicationControllersPage/>} />
+          <Route exact path='/replicationcontrollers/:uid' element={<rc.ReplicationControllersDetailPage/>} />
+          <Route exact path='/replicationcontrollers/:uid/edit' element={<rc.ReplicationControllersEditPage/>} />
+          <Route exact path='/roles' element={<roles.RolesPage/>} />
+          <Route exact path='/roles/:uid' element={<roles.RolesDetailPage/>} />
+          <Route exact path='/roles/:uid/edit' element={<roles.RolesEditPage/>} />
+          <Route exact path='/routes' element={<routes.RoutesPage/>} />
+          <Route exact path='/routes/:uid' element={<routes.RoutesDetailPage/>} />
+          <Route exact path='/routes/:uid/edit' element={<routes.RoutesEditPage/>} />
+          <Route exact path='/search' element={<search.SearchPage/>} />
+          <Route exact path='/secrets' element={<secrets.SecretsPage/>} />
+          <Route exact path='/secrets/:uid' element={<secrets.SecretsDetailPage/>} />
+          <Route exact path='/secrets/:uid/edit' element={<secrets.SecretsEditPage/>} />
+          <Route exact path='/services' element={<services.ServicesPage/>} />
+          <Route exact path='/services/:uid' element={<services.ServicesDetailPage/>} />
+          <Route exact path='/services/:uid/edit' element={<services.ServicesEditPage/>} />
+          <Route exact path='/statefulsets' element={<sts.StatefulSetsPage/>} />
+          <Route exact path='/statefulsets/:uid' element={<sts.StatefulSetsDetailPage/>} />
+          <Route exact path='/statefulsets/:uid/edit' element={<sts.StatefulSetsEditPage/>} />
+          {/*<Route element={<Error404Page/>} />*/}
+        </Routes>
       </Router>
   );
 }
