@@ -15,12 +15,13 @@
  *
  */
 import React from 'react';
+import {withParams} from '../router';
 import md from '../metadata';
 import crb from './';
 import ResourceEditPage from '../components/ResourceEditPage';
 import Link from '../components/Link';
 
-const ClusterRoleBindingsEditPage = ({match: {params: {uid}}}) => (
+const ClusterRoleBindingsEditPage = ({params: {uid}}) => (
   <ResourceEditPage
     kind='ClusterRoleBindings'
     path='clusterrolebindings'
@@ -32,4 +33,4 @@ const ClusterRoleBindingsEditPage = ({match: {params: {uid}}}) => (
   />
 );
 
-export default ClusterRoleBindingsEditPage;
+export default withParams(ClusterRoleBindingsEditPage);
