@@ -19,6 +19,7 @@ import {connect} from 'react-redux';
 import {AutoSizer, List} from 'react-virtualized';
 import Convert from 'ansi-to-html';
 import dompurify from 'dompurify';
+import {withParams} from '../router';
 import metadata from '../metadata';
 import cnt from '../containers';
 import p from '../pods';
@@ -108,7 +109,7 @@ const PodsLogsPage = ({uid, namespace, name, containers}) => {
 
 const mapStateToProps = ({pods}) => ({pods});
 
-const mergeProps = ({pods}, dispatchProps, {match: {params: {uid}}}) => ({
+const mergeProps = ({pods}, dispatchProps, {params: {uid}}) => ({
   ...dispatchProps,
   uid,
   namespace: metadata.selectors.namespace(pods[uid]),
@@ -116,4 +117,4 @@ const mergeProps = ({pods}, dispatchProps, {match: {params: {uid}}}) => ({
   containers: p.selectors.containers(pods[uid])
 });
 
-export default connect(mapStateToProps, null, mergeProps)(PodsLogsPage);
+export default withParams(connect(mapStateToProps, null, mergeProps)(PodsLogsPage));
